Refactoring DeliveryOrderService. Created new classes  - FlightRecordUtils  - DroneLogisticsService
